pyproject.toml
[tool.poetry]
name = "dot-trees"
version = "0.1.0"
description = ""
authors = ["capricorn <8250820+capricorn@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
